GUACAMOLE-990: Guarantee auth banning occurs before all other auth processing.
