+++ extensions/guacamole-auth-ban/src/main/resources/guac-manifest.json
@@ -5,10 +5,6 @@
     "name"      : "Brute-force Authentication Detection/Prevention",
     "namespace" : "ban",
 
-    "authProviders" : [
-        "org.apache.guacamole.auth.ban.BanningAuthenticationProvider"
-    ],
-
     "listeners" : [
         "org.apache.guacamole.auth.ban.BanningAuthenticationListener"
     ],
